collector/requirements.txt
requests
psycopg2-binary
psutil
